2 - Decouple tests into different file and evals file rename
